refactor: norm対応
一行読む部分をread_object.cに切り分けた。

#43
