+++ includes/reader.h
@@ -9,5 +9,8 @@
 # include <errno.h>
 
 void	read_rtfile(t_data *data, char *filepath);
+void	read_circle(t_object *circle, FILE *file);
+void	read_triangle(t_object *triangle, FILE *file);
+void	exit_error(char *errmsg);
 
 #endif
